Move nested routes to BodyNav.js, edit css styling and rename Navbar, BodyNavBar
