+++ src/App.css
@@ -1,0 +1,3 @@
+.CCPC__main {
+  height: 100vh;
+}
