+++ src/components/navBar/Nav.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
 
-import "./Navbar.css";
+import "./Nav.css";
 
 const Navbar = ({ sermons, live, about }) => {
   return (
