+++ src/pages/contact/ContactUs.js
@@ -1,5 +1,6 @@
 import React, { Component } from "react";
 import "../../stylesheet/Styles.css";
+
 import "../contact/ContactUs.css";
 import Divider from "../../assets/Divider.png";
 
@@ -8,14 +9,14 @@ class Contact extends Component {
 
   render() {
     return (
-      <div className="container">
+      <>
         <h2 className="sub__title">Contact Us</h2>
         <img className="divider" src={Divider} alt="" />
         <div className="form__body content">
           <p>Helloooooo, call me please!</p>
           {/* <form></form> */}
         </div>
-      </div>
+      </>
     );
   }
 }
