+++ src/pages/kimBio/Bio.js
@@ -1,17 +1,13 @@
 import React from "react";
-import { useParams } from "react-router-dom";
 import "../../stylesheet/Styles.css";
 import "./Bio.css";
 import ProfilePic from "../../assets/TestProfile.png";
 import Divider from "../../assets/Divider.png";
 
 const Bio = () => {
-  // let { bioId } = useParams();
-  console.log("hello");
   return (
-    <div className="container">
+    <>
       <h2 className="sub__title">Rev. Ki Kim</h2>
-      {/* <h2>{bioId}</h2> */}
       <img className="divider" src={Divider} alt="" />
       <div className="bio__content content">
         <img className="bio__pic" src={ProfilePic} alt="" />
@@ -45,7 +41,7 @@ const Bio = () => {
           deserunt. Commodo minim aute labore culpa.
         </p>
       </div>
-    </div>
+    </>
   );
 };
