+++ src/stylesheet/Styles.css
@@ -14,6 +14,7 @@ code {
 /* CURRENT DIVIDER IMAGE CSS */
 .divider {
   margin-bottom: 25px;
+  /* width: 10%; */
 }
 
 /* STATIC PAGES CSS */
